Tests for Mixin classes (mixin.py) (#24)
* Tests for WorkbookMixinElement class

* Fixed some tests

* Update setUp method for WorkbookMixinElement tests

* Added excessive_parameters test

* Added tests for TopicMixinElement class

+++ xmind/core/mixin.py
@@ -43,7 +43,7 @@ class WorkbookMixinElement(Element):
             return utils.readable_time(timestamp)
 
     def setModifiedTime(self, time):
-        self.setAttribute(const.ATTR_TIMESTAMP, long(time))
+        self.setAttribute(const.ATTR_TIMESTAMP, int(time))
 
     def updateModifiedTime(self):
         self.setModifiedTime(utils.get_current_time())
@@ -53,6 +53,7 @@ class WorkbookMixinElement(Element):
 
 
 class TopicMixinElement(Element):
+    # FIXME: init input parameters is without default initialization
     def __init__(self, node, ownerTopic):
         super(TopicMixinElement, self).__init__(node)
         self._owner_topic = ownerTopic
